Mere with Voie

+++ Mere.cpp
@@ -26,6 +26,7 @@
 #include "Clavier.h"
 #include "Heure.h"
 #include "Feux.h"
+#include "Voie.h"
 
 ///////////////////////////////////////////////////////////////////  PRIVE
 //------------------------------------------------------------- Constantes
@@ -62,6 +63,11 @@ int main ( void )
 	pid_t pidC;
 	pid_t pidH;
 	pid_t pidF;
+	pid_t pidN;
+	pid_t pidS;
+	pid_t pidE;
+	pid_t pidO;
+	
 	InitialiserApplication(XTERM);
 
 	//bloquage de sigint
@@ -88,9 +94,10 @@ int main ( void )
 	//creation d'un boite aux lettres
 	int balVoitures = msgget(key, IPC_CREAT | IPC_EXCL | 0600);
 
+
 	if((pidC=fork())==0)
 	{
-		Clavier();
+		Clavier(semDureeFeuId, mpDureeFeuId, balVoitures);
 	}
 	else
 	{
@@ -100,32 +107,69 @@ int main ( void )
 		}
 		else
 		{
-			pidH = CreerEtActiverHeure();
-			
-			//on attend la terminaison du clavier (par q ou sigint)
-			waitpid(pidC, 0, 0);
-			//on envoie le signal de fin à la tache heure
-			kill(pidH, SIGUSR2);
-			waitpid(pidH, 0, 0);
-			//on envoie le signal de fin à la tache feux
-			kill(pidF, SIGUSR2);
-			waitpid(pidF, 0, 0);
+			if((pidN=fork())==0)
+			{
+				ActiverVoie(NORD, balVoitures, semCouleurFeuId, mpCouleurFeuId);	
+			}else
+			{
+				if((pidS=fork())==0)
+				{
+					ActiverVoie(SUD, balVoitures, semCouleurFeuId, mpCouleurFeuId);	
+				}else
+				{
+					if((pidE=fork())==0)
+					{
+						ActiverVoie(EST, balVoitures, semCouleurFeuId, mpCouleurFeuId);	
+					}else
+					{
+						if((pidO=fork())==0)
+						{
+							ActiverVoie(OUEST, balVoitures, semCouleurFeuId, mpCouleurFeuId);	
+						}else
+						{
+							pidH = CreerEtActiverHeure();
+							
 
-			//suppression de la B.A.L.
-			msgctl(balVoitures, IPC_RMID, 0);
+							//on attend la terminaison du clavier (par q ou sigint)
+							waitpid(pidC, 0, 0);
+							//on envoie le signal de fin à la tache heure
+							kill(pidH, SIGUSR2);
+							waitpid(pidH, 0, 0);
+							//on envoie le signal de fin à la tache feux
+							kill(pidF, SIGUSR2);
+							waitpid(pidF, 0, 0);
+							
+							//on envoie le signal de fin à la tache VoieN
+							kill(pidN, SIGUSR2);
+							waitpid(pidN, 0, 0);
+							//on envoie le signal de fin à la tache VoieS
+							kill(pidS, SIGUSR2);
+							waitpid(pidS, 0, 0);
+							//on envoie le signal de fin à la tache VoieE
+							kill(pidE, SIGUSR2);
+							waitpid(pidE, 0, 0);
+							//on envoie le signal de fin à la tache VoieO
+							kill(pidO, SIGUSR2);
+							waitpid(pidO, 0, 0);
 
-			//suppression des memoires partagees
-			shmctl(mpDureeFeuId, IPC_RMID, 0);
-			shmctl(mpCouleurFeuId, IPC_RMID, 0);
+							//suppression de la B.A.L.
+							msgctl(balVoitures, IPC_RMID, 0);
 
-			//supression des semaphores
-			semctl(semDureeFeuId, 0, IPC_RMID, 0);
-			semctl(semCouleurFeuId, 0, IPC_RMID, 0);
+							//suppression des memoires partagees
+							shmctl(mpDureeFeuId, IPC_RMID, 0);
+							shmctl(mpCouleurFeuId, IPC_RMID, 0);
 
-			TerminerApplication();
+							//supression des semaphores
+							semctl(semDureeFeuId, 0, IPC_RMID, 0);
+							semctl(semCouleurFeuId, 0, IPC_RMID, 0);
+
+							TerminerApplication();
+						}
+					}
+				}
+			}
 		}
 	}
 	//return ou exit ?
 	return 0;
 } //----- fin de Nom
-
